File with snippent to book mark js script
Utils, short for utilities, are tools or libraries that can be used to perform certain tasks or functions in JavaScript code. These utilities can be bookmarklets, browser extensions, or standalone libraries that can be added to a web page to enhance its functionality. With the help of these utilities, web developers can perform various tasks such as debugging, optimizing code, manipulating the DOM, and much more. These utilities can greatly improve the performance and functionality of a website page by simplifying complex tasks and allowing developers to focus on building more innovative features.